refactor test helper
