+++ packages/ts-compiler/src/typescript-compiler.ts
@@ -24,6 +24,7 @@ export class TypescriptCompiler implements Compiler {
   }
 
   public getDynamicPackageDependencies(ctx: CompilerContext, name?: string) {
+    debugger;
     const deps = this.preset.getDynamicPackageDependencies ? this.preset.getDynamicPackageDependencies() : {};
     return deps;
   }
